Restructure tests, add eslint
