+++ .eslintrc
@@ -1,0 +1,49 @@
+{
+    "env": {
+        "browser": true,
+        "node": true
+    },
+    "parser": "babel-eslint",
+    "plugins": [
+        "react"
+    ],
+    "ecmaFeatures": {
+        "arrowFunctions": true,
+        "blockBindings": true,
+        "destructuring": true,
+        "jsx": true,
+        "modules": true,
+        "objectLiteralComputedProperties": true,
+        "objectLiteralDuplicateProperties": true,
+        "objectLiteralShorthandMethods": true,
+        "objectLiteralShorthandProperties": true,
+        "spread": true
+    },
+    "rules": {
+        "camelcase": 0,
+        "eol-last": 0,
+        "new-cap": 0,
+        "quotes": 0,
+        "semi": 1,
+        "no-unused-vars": 2,
+        "no-underscore-dangle": 0,
+        "no-unused-expressions": 0,
+        "no-undef": 2,
+
+        "react/jsx-no-undef": 2,
+        "react/jsx-uses-react": 1,
+        "react/jsx-uses-vars": 1,
+        "react/no-did-mount-set-state": 2,
+        "react/no-did-update-set-state": 2,
+        "react/no-multi-comp": 1,
+        "react/no-unknown-property": 1,
+        "react/react-in-jsx-scope": 1,
+        "react/self-closing-comp": 1,
+        "react/prop-types": [
+            1,
+            {
+                "ignore": ["children"]
+            }
+        ]
+    },
+}
