+++ package.json
@@ -58,8 +58,11 @@
     "react": "^0.13.2"
   },
   "devDependencies": {
+    "babel-eslint": "^5.0.0",
     "babel-loader": "^6.2.1",
     "chai": "^3.2.0",
+    "eslint": "^2.2.0",
+    "eslint-plugin-react": "^4.1.0",
     "mocha": "^2.3.0",
     "node-inspector": "^0.12.3",
     "nodeunit": "^0.9.0",
