+++ test/.eslintrc
@@ -1,0 +1,5 @@
+{
+    "env": {
+        "mocha": true
+    }
+}
